cuda kernel launch error (sac)

+++ src/rl/environments/pendulum/sac/cpu/training.cpp
@@ -15,7 +15,7 @@
 namespace rlt = rl_tools;
 
 using DEVICE = rlt::devices::DEVICE_FACTORY<>;
-using T = float;
+using T = double;
 using TI = typename DEVICE::index_t;
 
 using PENDULUM_SPEC = rlt::rl::environments::pendulum::Specification<T, TI, rlt::rl::environments::pendulum::DefaultParameters<T>>;
